requirements.txt
fastmcp
pywinauto
pywin32
uiautomation
mss
Pillow